requirements.txt
ebooklib>=0.18
pillow>=10.0.0
PyPDF2>=3.0.0
requests>=2.31.0